Put Server in its file
